refactor: Abstracted and combined jobDescriptions into types folder
